backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.2.0
pydantic==2.12.5
pydantic-settings==2.13.0
email-validator==2.0.0
python-multipart==0.0.7
redis==7.4.0
bcrypt==4.0.1
cloudinary==1.32.0
PyJWT==2.8.0
python-slugify==8.0.1
reportlab==4.0.0
httpx==0.27.0
